Updated Relativistic Profile (#7)
Tidied up the relativistic emission profile, using UTF-8 formatting where appropriate. Added comments to functions, and added labels to plots. Renamed the file for clarity.

Fairly sure that the shape we have so far looks about right, however the scale of the temperature plot is very far away from what we are expecting. Not sure yet what the issue is, but we are looking at the non-relativistic version first to see if the issue persists and will work back from there.