Moved and Refactored Temperature Render, and Added Animations
Moved all test scripts to a new scripts folder. Refactored the code in `temperature_render.jl` to be in a function, with default arguments that can be quickly changed to modify the properties of the render.

This is used with a test animation script which can generate animations, varying parameters such as the observation angle.