Small Changes To Renders
Made a couple of changes to the title of the plots to make them a bit clearer, and added a quick commentable option for contour maps.

Added a frame counter to `animation_test.jl` to see the progress of the render.

Removed the `NR_Emission_Profile .jl` from `src`, as it doesn't seem necessary anymore, but has been backed up.