Added Comments for Potential Units Fixes
We went completely went through the units together, and I think we have found exactly how they should be done to give correct units. No changes have been made, so that it doesn't break anything, but the way things should be done, or any questions have been commented in.

Here's what we gathered:

-`r_isco` should be in units of M, and , and by implementing line `10`, we should have the same values from AccretionFormulae and CarterBoyerLindquist.

-The `mdot` function seems to be where our problems are coming from, as the `M` going into line `47` is using the units from CarterBoyerLindquist will just be 1, where it should be something large in kg, as that is what the `M_\sun` is in (or convert `M_\sun the other way). We aren't certain what the units are, but we recall a c^2 factor being mentioned to convert it to kg.

-The `diss` function's output currently has dimensions 1/MT, but physically represents flux which should have dimensions M/T^3. To correct for this, we found a combination of c and G which give the required dimensions M^2/T^2, which was c^6/G^2, this should give us a physical temperature.

If everything we worked out here is correct, then we should be able to input all of these changes, and it should give us more physical results.