Possible Fixes For Units (#21)
* Fixed merge issues

Previous commit, hopefully not overriding changes to `redshift.jl`

* Removed Plots import.

Removed Plots import from emission_profile.jl

* Moved and Refactored Temperature Render, and Added Animations

Moved all test scripts to a new scripts folder. Refactored the code in `temperature_render.jl` to be in a function, with default arguments that can be quickly changed to modify the properties of the render.

This is used with a test animation script which can generate animations, varying parameters such as the observation angle.

* Small Changes To Renders

Made a couple of changes to the title of the plots to make them a bit clearer, and added a quick commentable option for contour maps.

Added a frame counter to `animation_test.jl` to see the progress of the render.

Removed the `NR_Emission_Profile .jl` from `src`, as it doesn't seem necessary anymore, but has been backed up.

* Added Comments for Potential Units Fixes

We went completely went through the units together, and I think we have found exactly how they should be done to give correct units. No changes have been made, so that it doesn't break anything, but the way things should be done, or any questions have been commented in.

Here's what we gathered:

-`r_isco` should be in units of M, and , and by implementing line `10`, we should have the same values from AccretionFormulae and CarterBoyerLindquist.

-The `mdot` function seems to be where our problems are coming from, as the `M` going into line `47` is using the units from CarterBoyerLindquist will just be 1, where it should be something large in kg, as that is what the `M_\sun` is in (or convert `M_\sun the other way). We aren't certain what the units are, but we recall a c^2 factor being mentioned to convert it to kg.

-The `diss` function's output currently has dimensions 1/MT, but physically represents flux which should have dimensions M/T^3. To correct for this, we found a combination of c and G which give the required dimensions M^2/T^2, which was c^6/G^2, this should give us a physical temperature.

If everything we worked out here is correct, then we should be able to input all of these changes, and it should give us more physical results.