Fixed Units and Mass Scaling
Corrected the the temperature that we got from a 1kg black hole, and scaled it up to an arbitrary mass black hole. This mass is measured in solar masses, and is calculated after the initial image is generated.

Also made some changes to the units which fixed some of the other scaling issues, so everything should all be correct now.

Removed a few old comments, and added a few `kwargs` to the `temperature_render()` function.

Also added an optional autoscaler which is commented out at the moment, so that the scale stays consistent for different masses.