Added Working f Function Approximations
Added a fully working integration of the f function using the numerical
approximation method that Fergus helped devlop, and with the equation
from Johannsen (2014). Allows you to change the values of epsilon, and
the alphas, which have an effect on the f function, which will change
the renders. The iron line plots are roughly the same, even with large
changes which is quite an interesting observations to note.